+++ src/process-file.js
@@ -5,16 +5,33 @@ const handleImportFailure = (err, file, st) => {
   st.logger.error(
     `Failed loading File: ${toRelativePath(file)} - ${err.message}`
   );
-  return {};
+  return [];
 };
 
-export default async (file, st) => {
-  const res = [];
+export const processFile = async (file, st) => {
   st.logger.debug(`Processing ${toRelativePath(file)}`);
-  let module = await import(file).catch((e) =>
-    handleImportFailure(e, file, st)
-  );
 
+  return import(file)
+    .then((module) => {
+      const res = [];
+      if (module.default !== undefined) {
+        module = module.default;
+      }
+      Object.keys(module).forEach((k) => {
+        const check = schema.checkSchema(module[k]);
+        if (check.error === undefined) {
+          st.logger.debug(`Importing Route: ${k}`);
+          res.push(module[k]);
+        }
+      });
+
+      return res;
+    }).
+    catch((e) => handleImportFailure(e, file, st));
+};
+
+export const processObject = (module, st) => {
+  const res = [];
   if (module.default !== undefined) {
     module = module.default;
   }
@@ -28,3 +45,5 @@ export default async (file, st) => {
 
   return res;
 };
+
+export default {};